+++ entities/systems/trolley.py
@@ -52,10 +52,10 @@ class Trolley(BaseSystem):
 
     def reset(self) -> None:
         """Reset: position, velocity, and delta_position to zero"""
-        self.position = torch.tensor(0, dtype=torch.float32)
-        self.velocity = torch.tensor(0, dtype=torch.float32)
-        self.delta_position = torch.tensor(0, dtype=torch.float32)
-        self.acceleration = torch.tensor(0, dtype=torch.float32)
+        self.position = torch.tensor(0)
+        self.velocity = torch.tensor(0)
+        self.delta_position = torch.tensor(0)
+        self.acceleration = torch.tensor(0)
 
     @property
     def X(self) -> Tensor:
